+++ NEWS.md
@@ -1,5 +1,8 @@
 # mlr3 (development version)
 
+* feat(Learner): Add support for a `$check_task()` method that can be
+                  implemented by `Learner`s to perform additional compatibility checks
+
 # mlr3 0.17.2
 
 * Skip new `data.table` tests on mac.